requirements.txt
orjson
curl-cffi
rich
tenacity
tqdm
nest-asyncio
pathlib
typing-extensions
google-cloud-storage
google-cloud-bigquery
python-dotenv
aiohttp
aiogoogle
tabulate
pandas
pyarrow
gcloud-aio-storage
aiofiles
google-cloud-logging